copied the content from the old app to the new app
