--- src/pages/about/index.astro
@@ -1,21 +1,0 @@
----
-import Layout from 'src/layouts/Layout.astro';
----
-
-<Layout title="About">
-	<main>
-		<h1> About </h1>
-		<p>
-			This is the about page
-		</p>
-		<p class="red">
-			red
-		</p>
-	</main>
-</Layout>
-
-<style>
-	.red {
-		color: red;
-	}
-</style>
