+++ src/pages/projects/index.astro
@@ -1,0 +1,20 @@
+---
+import Layout from 'src/layouts/Layout.astro';
+---
+
+<Layout title="Experimental Projects">
+	<main>
+		<h1> Experimental Projects </h1>
+		<p> 
+			These are some of the other projects I have worked on. 
+			Most of these projects are experiments or prototypes. None the less, 
+			I am cool with anyone looking at these projects.
+		</p>
+		<p><span class="project_title"> Calculator Desktop App (Rust, GTK4, Git) </span> <br />
+			Git Repository: <a href="https://github.com/sagedemage/CalculatorDesktopApp"> CalculatorDesktopApp </a>
+		</p>
+		<ul>
+			<li> Designed the UI with accessibility in mind </li>
+		</ul>
+	</main>
+</Layout>
