+++ tsconfig.json
@@ -2,7 +2,9 @@
   "compilerOptions": {
     "baseUrl": "./",
     "jsx": "react-jsx",
-    "jsxImportSource": "react"
+    "jsxImportSource": "react",
   },
+  "include": ["env.d.ts", "**/*.ts", "**/*.tsx", "./"],
+  "exclude": ["node_modules"],
   "extends": "astro/tsconfigs/strictest"
 }
